prefer-step-decorator lint, AxeCacheHelper

+++ playwright/global/teardown-local.ts
@@ -3,7 +3,10 @@ import urls from '../config/urls';
 import UserType from '../enums/user-type';
 import UserStateHelper from '../helpers/users-state-helper';
 import FileSystemHelper from '../helpers/file-system-helper';
+import config from '../config/config';
 import { request } from 'playwright-core';
+import AxeCacheHelper from '../helpers/axe-cache-helper';
+import CookiesHelper from '../helpers/cookies-helper';
 
 //This is last resort teardown for citizen users if test execution gets interupted in local.
 
@@ -28,10 +31,13 @@ const deleteUsers = async (userType: UserType) => {
 };
 
 const globalTeardownLocal = async () => {
-  await deleteUsers(UserType.CLAIMANT);
-  await deleteUsers(UserType.DEFENDANT);
+  if (!config.skipCitizenSetup) {
+    await deleteUsers(UserType.CLAIMANT);
+    await deleteUsers(UserType.DEFENDANT);
+  }
   UserStateHelper.deleteAllUsersState();
-  FileSystemHelper.delete(`${filePaths.userCookies}/`);
+  CookiesHelper.deleteAllCookies();
+  AxeCacheHelper.deleteAllCache();
 };
 
 export default globalTeardownLocal;
